requirements.txt
SoundFile==0.10.3.post1
ipdb==0.13.4
pypesq==1.2.4
scipy==1.5.4
h5py==2.10.0
torchaudio==0.7.0
apex==0.9.10.dev0
torch==1.7.0
webrtcvad==2.0.10
librosa==0.6.2
gdown==3.12.2
tensorflow
matplotlib==3.3.3
tqdm==4.54.0
multiprocess==0.70.12.2
numpy==1.21.1
parallel_wavegan==0.4.8
PyYAML==5.4.1
scikit_learn==0.24.2
tensorboardX
umap==0.1.1
visdom==0.1.8.9
